65l post going[65%]...

+++ internal/service/post.go
@@ -27,14 +27,14 @@ func (s *Post) Create(data *dtos.PostFile, userId int) (*dtos.Post, error) {
 	var err error
 	data.PhotoPath, err = utill.Upload(data.Photo, config.PostDir)
 	if err != nil {
-		return nil, fmt.Errorf("error uploading photo: ", err.Error())
+		return nil, fmt.Errorf("error uploading photo: %s", err.Error())
 	}
 
 	data.MediaType = models.Photo
 	if data.Video != nil {
 		data.VideoPath, err = utill.Upload(data.Video, config.PostDir)
 		if err != nil {
-			return nil, fmt.Errorf("error uploading video: ", err.Error())
+			return nil, fmt.Errorf("error uploading video: %s", err.Error())
 		}
 		data.MediaType = models.Video
 	}
@@ -51,7 +51,7 @@ func (s *Post) Create(data *dtos.PostFile, userId int) (*dtos.Post, error) {
 
 	res := new(dtos.Post)
 	res.MapFromModel(model)
-	res.Photo = utill.PutMediaDomain(res.Photo)
+	res.Photo = utill.PutMediaPostDomain(res.Photo)
 
 	return res, nil
 }
@@ -71,17 +71,95 @@ func (s *Post) GetList(data *dtos.PostFilter) ([]*dtos.Post, error) {
 		post := new(dtos.Post).
 			MapFromModel(model)
 
-		post.Photo = utill.PutMediaDomain(post.Photo)
+		post.Photo = utill.PutMediaPostDomain(post.Photo)
 		listDto[i] = post
 	}
 
 	return listDto, nil
 }
 
-func (s *Post) GetById(id int) (*dtos.PostDetail, error) {
-	return s.Repo.Post.GetDetail(id)
+func (s *Post) GetDetail(id int) (*dtos.PostDetail, error) {
+	dto, err := s.Repo.Post.GetDetail(id)
+	if err != nil {
+		return nil, err
+	}
+
+	dto.PhotoPath = utill.PutMediaPostDomain(dto.PhotoPath)
+	dto.PosterPhotoUrl = utill.PutMediaDomain(dto.PosterPhotoUrl)
+
+	return dto, nil
 }
 
-func (s *Post) DeleteById(id int) error {
+func (s *Post) Update(data *dtos.PostFile, userId int) (*dtos.Post, error) {
+	if !s.Repo.Business.ExistsByIdAndUserId(data.BusinessId, userId) {
+		return nil, errors.New("permission denied: you are not owner of this business profile")
+	}
+	model, err := s.Repo.Post.GetById(data.Id)
+	if err != nil {
+		return nil, err
+	}
+
+	if model.BusinessId != data.BusinessId {
+		return nil, errors.New("access denied")
+	}
+
+	if data.MediaType == "" {
+		if data.Video == nil && model.VideoPath == "" {
+			data.MediaType = models.Photo
+		} else {
+			data.MediaType = models.Video
+		}
+	}
+
+	switch data.MediaType {
+	case models.Photo:
+		model.VideoPath = ""
+		model.MediaType = models.Photo
+	case models.Video:
+		if data.Video == nil || model.VideoPath == "" {
+			return nil, errors.New("video not found")
+		}
+		model.MediaType = models.Video
+	default:
+		return nil, errors.New("media_type not found")
+	}
+
+	if data.Photo != nil {
+		model.PhotoPath, err = utill.Upload(data.Photo, config.PostDir)
+		if err != nil {
+			return nil, fmt.Errorf("error uploading photo: %s", err.Error())
+		}
+	}
+
+	if data.Video != nil {
+		data.VideoPath, err = utill.Upload(data.Video, config.PostDir)
+		if err != nil {
+			return nil, fmt.Errorf("error uploading video: %s", err.Error())
+		}
+		model.MediaType = models.Video
+	}
+
+	model, err = s.Repo.Post.Update(model)
+	if err != nil {
+		return nil, err
+	}
+
+	res := new(dtos.Post)
+	res.MapFromModel(model)
+	res.Photo = utill.PutMediaPostDomain(res.Photo)
+
+	return res, nil
+}
+
+func (s *Post) DeleteById(id int, userId int) error {
+	model, err := s.Repo.Post.GetById(id)
+	if err != nil {
+		return err
+	}
+
+	if !s.Repo.Business.ExistsByIdAndUserId(model.BusinessId, userId) {
+		return errors.New("access denied")
+	}
+
 	return s.Repo.Post.DeleteById(id)
 }
